Time: 2 ms (97.18%) | Memory: 46 MB (27.44%) - LeetSync

+++ 637-average-of-levels-in-binary-tree/average-of-levels-in-binary-tree.java
@@ -37,7 +37,7 @@ class Solution {
                     q.add(node.right);
                 }
             }
-            double average = sum / size;
+            double average = sum / (double)size;
             res.add(average);
         }
